Rename some classes
Minor refactoring
