+++ NFluent.35/Extensibility/FluentMessage.cs
@@ -17,7 +17,6 @@ namespace NFluent.Extensibility
 {
     using System;
     using System.Text;
-    using NFluent.Extensions;
 
     // TODO: probably worth to refactor the implementation of this class
 
@@ -287,229 +286,11 @@ namespace NFluent.Extensibility
         /// </summary>
         /// <param name="value">The value to get the type from.</param>
         /// <returns>The appropriate entity label.</returns>
-        private string GetEntityFromType(object value)
+        internal string GetEntityFromType(object value)
         {
             return this.entity ?? DefaultEntity;
         }
 
         #endregion
-
-        /// <summary>
-        /// Class describing a message block.
-        /// </summary>
-        public class MessageBlock
-        {
-            #region fields
-
-            private readonly FluentMessage message;
-
-            private readonly IValueBlock value;
-
-            private readonly object test;
-
-            private readonly string attribute;
-
-            private string customMessage;
-
-            private string comparisonLabel;
-
-            #endregion
-
-            #region constructor
-
-            /// <summary>
-            /// Initializes a new instance of the <see cref="MessageBlock"/> class.
-            /// </summary>
-            /// <param name="message">The message.</param>
-            /// <param name="test">The tested object.</param>
-            /// <param name="attribute">The block attribute.</param>
-            public MessageBlock(FluentMessage message, object test, string attribute) : this(message, test.GetTypeWithoutThrowingException(), attribute)
-            {
-                this.test = test;
-                this.value = new ValueBlock(test);
-            }
-
-            /// <summary>
-            /// Initializes a new instance of the <see cref="MessageBlock"/> class.
-            /// </summary>
-            /// <param name="message">The message.</param>
-            /// <param name="type">The tested type.</param>
-            /// <param name="attribute">The block attribute.</param>
-            public MessageBlock(FluentMessage message, Type type, string attribute)
-            {
-                if (message == null)
-                {
-                    throw new ArgumentNullException("message");
-                }
-
-                this.value = new InstanceBlock(type);
-                this.message = message;
-                this.test = null;
-                this.attribute = attribute;
-            }
-            #endregion
-
-            #region properties
-
-           /// <summary>
-           /// Gets the Message.
-           /// </summary>
-           /// <value>
-           /// The <see cref="FluentMessage"/> holding that block.
-           /// </value>
-           public FluentMessage And
-           {
-               get
-               {
-                   return this.message;
-               }
-           }
-
-           #endregion
-
-            #region methods
-
-           /// <summary>
-           /// Gets the message as a string.
-           /// </summary>
-           /// <returns>A string with the properly formatted message.</returns>
-           public string GetMessage()
-           {
-               var builder = new StringBuilder();
-               builder.Append(this.FullLabel());
-               builder.Append(EndOfLine);
-               builder.Append(this.value.GetMessage());
-               return builder.ToString();
-           }
-
-           private string FullLabel()
-           {
-               string fullLabel;
-               if (string.IsNullOrEmpty(this.comparisonLabel))
-               {
-                   fullLabel = string.Format(
-                       this.customMessage ?? "The {0} {1}:",
-                       this.attribute,
-                       this.message.GetEntityFromType(this.test));
-               }
-               else
-               {
-                   fullLabel = string.Format(
-                       this.customMessage ?? "The {0} {1}: {2}",
-                       this.attribute,
-                       this.message.GetEntityFromType(this.test),
-                       this.comparisonLabel);
-               }
-
-               return fullLabel;
-           }
-
-           /// <summary>
-           /// Gets the block label.
-           /// </summary>
-           /// <returns>The block Label.</returns>
-           public string GetBlockLabel()
-           {
-               return string.Format(
-                   "{0} {1}", this.attribute, this.message.GetEntityFromType(this.test));
-           }
-
-           /// <summary>
-           /// Adds a specific comparison message (e.g 'equal to').
-           /// </summary>
-           /// <param name="comparison">
-           /// The comparison suffix.
-           /// </param>
-           /// <returns>
-           /// The <see cref="MessageBlock"/> for fluent API.
-           /// </returns>
-           public MessageBlock Comparison(string comparison)
-           {
-               this.comparisonLabel = comparison;
-               return this;
-           }
-
-           /// <summary>
-           /// Returns a <see cref="System.String" /> that represents this instance.
-           /// </summary>
-           /// <returns>
-           /// A <see cref="System.String" /> that represents this instance.
-           /// </returns>
-           public override string ToString()
-           {
-               return this.message.ToString();
-           }
-
-           /// <summary>
-           /// Requests that the Hash value is included in the description block.
-           /// </summary>
-           /// <param name="active">
-           /// True to include the type. This is the default value.
-           /// </param>
-           /// <returns>
-           /// Returns this instance for chained calls.
-           /// </returns>
-           public MessageBlock WithHashCode(bool active = true)
-           {
-               this.value.WithHashCode(active);
-               return this;
-           }
-
-           /// <summary>
-           /// Requests that the type is included in the description block.
-           /// </summary>
-           /// <param name="active">
-           /// True to include the type. This is the default value.
-           /// </param>
-           /// <param name="full">
-           /// True to display the full type name (with assembly).
-           /// </param>
-           /// <returns>
-           /// Returns this instance for chained calls.
-           /// </returns>
-           public MessageBlock WithType(bool active = true, bool full = false)
-           {
-               this.value.WithType(active, full);
-               return this;
-           }
-
-           /// <summary>
-           /// Requests that a specific type is included in the description block.
-           /// </summary>
-           /// <param name="forcedType">Type to include in the description.</param>
-           /// <remarks>Default type is the type of the object instance given in constructor.</remarks>
-           /// <returns>
-           /// Returns this instance for chained calls.
-           /// </returns>
-           public MessageBlock WithType(Type forcedType)
-           {
-               this.value.WithType(forcedType);
-               return this;
-           }
-
-           /// <summary>
-           /// Specifies a specific attribute for the message.
-           /// </summary>
-           /// <param name="newLabel">The new attribute.</param>
-           /// <returns>This <see cref="MessageBlock"/>.</returns>
-           public MessageBlock Label(string newLabel)
-           {
-               this.customMessage = newLabel;
-               return this;
-           }
-
-           /// <summary>
-           /// Adds a description of the number of items (only relevant if the object is an enumerable).
-           /// </summary>
-           /// <param name="itemsCount">The number of items of the enumerable instance.</param>
-           /// <returns>The description of the number of items (only relevant if the object is an enumerable).</returns>
-           public MessageBlock WithEnumerableCount(long itemsCount)
-           {
-               this.value.WithEnumerableCount(itemsCount);
-               return this;
-           }
-
-           #endregion
-        }
     }
 }
