+++ NFluent.35/NFluent.35.csproj
@@ -64,7 +64,8 @@
     <Compile Include="Assertions\EventWaitHandleCheckExtensions.cs" />
     <Compile Include="Assertions\ObjectFieldsCheckExtensions.cs" />
     <Compile Include="Extensibility\InstanceBlock.cs" />
-    <Compile Include="Extensibility\IValueBlock.cs" />
+    <Compile Include="Extensibility\IValueDescription.cs" />
+    <Compile Include="Extensibility\MessageBlock.cs" />
     <Compile Include="Extensibility\ValueBlock.cs" />
     <Compile Include="FluentCodeCheck.cs" />
     <Compile Include="Extensibility\ExtensibilityHelper.cs" />
