+++ NFluent.40/NFluent.40.csproj
@@ -152,8 +152,11 @@
     <Compile Include="..\NFluent.35\Extensibility\InstanceBlock.cs">
       <Link>Extensibility\InstanceBlock.cs</Link>
     </Compile>
-    <Compile Include="..\NFluent.35\Extensibility\IValueBlock.cs">
-      <Link>Extensibility\IValueBlock.cs</Link>
+    <Compile Include="..\NFluent.35\Extensibility\IValueDescription.cs">
+      <Link>Extensibility\IValueDescription.cs</Link>
+    </Compile>
+    <Compile Include="..\NFluent.35\Extensibility\MessageBlock.cs">
+      <Link>Extensibility\MessageBlock.cs</Link>
     </Compile>
     <Compile Include="..\NFluent.35\Extensibility\ValueBlock.cs">
       <Link>Extensibility\ValueBlock.cs</Link>
@@ -363,7 +366,7 @@
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <ProjectExtensions>
     <VisualStudio>
-      <UserProperties ProjectLinkReference="3a474510-f384-41ce-9270-539b1e51da7e" ProjectLinkerExcludeFilter="\\?desktop(\\.*)?$;\\?silverlight(\\.*)?$;\.desktop;\.silverlight;\.xaml;^service references(\\.*)?$;\.clientconfig;^web references(\\.*)?$" />
+      <UserProperties ProjectLinkerExcludeFilter="\\?desktop(\\.*)?$;\\?silverlight(\\.*)?$;\.desktop;\.silverlight;\.xaml;^service references(\\.*)?$;\.clientconfig;^web references(\\.*)?$" ProjectLinkReference="3a474510-f384-41ce-9270-539b1e51da7e" />
     </VisualStudio>
   </ProjectExtensions>
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
