+++ NFluent.PCL/NFluent.PCL.csproj
@@ -157,8 +157,11 @@
     <Compile Include="..\NFluent.35\Extensibility\InstanceBlock.cs">
       <Link>Extensibility\InstanceBlock.cs</Link>
     </Compile>
-    <Compile Include="..\NFluent.35\Extensibility\IValueBlock.cs">
-      <Link>Extensibility\IValueBlock.cs</Link>
+    <Compile Include="..\NFluent.35\Extensibility\IValueDescription.cs">
+      <Link>Extensibility\IValueDescription.cs</Link>
+    </Compile>
+    <Compile Include="..\NFluent.35\Extensibility\MessageBlock.cs">
+      <Link>Extensibility\MessageBlock.cs</Link>
     </Compile>
     <Compile Include="..\NFluent.35\Extensibility\ValueBlock.cs">
       <Link>Extensibility\ValueBlock.cs</Link>
@@ -349,7 +352,7 @@
   <Import Project="$(MSBuildExtensionsPath32)\Microsoft\Portable\$(TargetFrameworkVersion)\Microsoft.Portable.CSharp.targets" />
   <ProjectExtensions>
     <VisualStudio>
-      <UserProperties ProjectLinkReference="3a474510-f384-41ce-9270-539b1e51da7e" ProjectLinkerExcludeFilter="\\?desktop(\\.*)?$;\\?silverlight(\\.*)?$;\.desktop;\.silverlight;\.xaml;^service references(\\.*)?$;\.clientconfig;^web references(\\.*)?$" />
+      <UserProperties ProjectLinkerExcludeFilter="\\?desktop(\\.*)?$;\\?silverlight(\\.*)?$;\.desktop;\.silverlight;\.xaml;^service references(\\.*)?$;\.clientconfig;^web references(\\.*)?$" ProjectLinkReference="3a474510-f384-41ce-9270-539b1e51da7e" />
     </VisualStudio>
   </ProjectExtensions>
   <Import Project="$(SolutionDir)\.nuget\NuGet.targets" Condition="Exists('$(SolutionDir)\.nuget\NuGet.targets')" />
